+++ Assets/Cookie Clicker/Runtime/Cookies/Domain/Baker/ICookieView.cs
@@ -4,10 +4,10 @@ namespace Cookie_Clicker.Runtime.Cookies.Domain.Baker
 {
     public interface ICookieView
     {
-        void UpdateStats(float totalCookies, float cps);
+        void UpdateStats(double totalCookies, double cps);
         void AddCursors(int amount);
         void RemoveCursors(int amount);
-        void Tap(float cookiesAmount);
+        void Tap(double cookiesAmount);
         
         void RegisterTapListener(Action listener);
     }
