+++ Assets/Cookie Clicker/Runtime/Cookies/Domain/Buildings/IBuildingStoreView.cs
@@ -8,7 +8,7 @@ namespace Cookie_Clicker.Runtime.Cookies.Domain.Buildings
         void Setup(List<BuildingDisplayData> displayDataList);
         void UpdateButtonsData(List<BuildingDisplayData> buildingsData);
         void UpdateButtonData(string buildingName, BuildingDisplayData displayData);
-        void UpdateButtonsInteraction(float currentCookies, PurchaseMode.Type purchaseType);
+        void UpdateButtonsInteraction(double currentCookies, PurchaseMode.Type purchaseType);
         void UpdateVisibility(string buildingName, BuildingVisibility visibility);
 
         void RegisterPurchasedModeListener(Action<PurchaseMode> listener);
